refactor: 추상 클래스 관련 테스트 통합하기.
